+++ pythonmysequel/Row.py
@@ -1,17 +1,22 @@
+'''
+This represents a record/row in a table
+It is mostly used as a parameter or attribute in
+pythonmysequel.Table and pythonmysequel.Connection
+'''
+
 from .Table import Table
 from .values import *
 
 class Row:
     def __init__(self,
-        table:Table,
-        **values
-    ) -> None:
+                table:Table, **values) -> None:
         self.table = table
         self.values = values
 
         self._check_value_type()
 
     def _check_value_type(self):
+        '''Make sure row values conform to table columns'''
         table_value = self.table.values
         for key, value, column_name in zip(self.values.keys(), self.values.values(), self.values):
             python_value_type = table_value[key].PYTHON_TYPE
@@ -21,7 +26,8 @@ class Row:
                 raise ValueError(f'Incorrect value type {inputted_value_type} for column "{column_name}" {python_value_type}')
 
 
-    def _add_value(self, value:dict):
+    def _add_value(self, value:dict[str: _ValueType]):
+        '''Add column data'''
         for k, v in value.items():
             self.values[k] = v